src/backend/database_manager/requirements.txt
fastapi
psycopg2-binary
python-dotenv
cryptography
pytest
pytest-asyncio
httpx
